Cleaned up code, added a nav bar to base, Fixed some variable fields of the model.

+++ budgetProject/budget/forms.py
@@ -2,6 +2,5 @@ from django import forms
 
 class ExpenseForm(forms.Form):
     title = forms.CharField()
-    amount = forms.IntegerField()
+    amount = forms.DecimalField()
     category = forms.CharField()
-
